refactor: move ShopHero component to section/hero for better structure
