+++ src/app/shop/_components/section/hero/ShopHero.tsx
